Module Dune_rpc.V1

module Id : sig ... end

Id's for requests, responses, sessions.

module Response : sig ... end

module Initialize : sig ... end

module Loc : sig ... end

module Path : sig ... end

module Ansi_color : sig ... end

module User_message : sig ... end

module Target : sig ... end

module Diagnostic : sig ... end

module Progress : sig ... end

module Job : sig ... end

module Sub : sig ... end

module Message : sig ... end

module Version_error : sig ... end

A Version_error is returned on the client-side when a request or notification is determined to be invalid due to version negotiation (no known method or no common version).

module Notification : sig ... end

module Request : sig ... end

module Client : sig ... end

module Where : sig ... end

represents the address where a dune rpc instance might be listening

module Registry : sig ... end

The registry is where all running instances of dune rpc are stored.