Module Ansi_color.RGB24

type t

24-bit RGB color (true color)

val red : t -> int

red t returns the red component of the 24-bit color t.

val green : t -> int

green t returns the green component of the 24-bit color t.

val blue : t -> int

blue t returns the blue component of the 24-bit color t.

val to_int : t -> int

to_int t returns the 24-bit color as an integer in the range 0, 0xFFFFFF. Each color components consists of 8 bits.