Module V1.Registry

The registry is where all running instances of dune rpc are stored.

It's used by clients to determine which dune rpc instance corresponds to the workspace they're trying to edit.

module Dune : sig ... end

a registered instance of dune. supposedly running and listening to rpc connections

module Config : sig ... end

The registry directory is located using xdg

type t

val create : Config.t -> t

val current : t -> Dune.t list

currently detected running instances

module Refresh : sig ... end

the result of polling the registry

module Poll (Fiber : sig ... end) (_ : sig ... end) : sig ... end

we can poll the registry efficiently using the following functor