Module V1.Request

type ('a, 'b) t

val ping : (unit, unit) t

val diagnostics : (unit, Diagnostic.t list) t

val format_dune_file : (Path.t * [ `Contents of string ], string) t

format a dune, dune-project, or a dune-workspace file. The full path to the file is necessary so that dune knows the formatting options for the project this file is in

val promote : (Path.t, unit) t

Promote a file.

val build_dir : (unit, Path.t) t

Returns the location of the build directory for the current build.