Module V1.Where

represents the address where a dune rpc instance might be listening

type t = [

| `Unix of string

| `Ip of [ `Host of string ] * [ `Port of int ]

]

type error =

| Invalid_where of string

exception E of error

module type S = sig ... end

module Make (Fiber : sig ... end) (_ : sig ... end) : S with type 'a fiber := 'a Fiber.t

obtain the address from the build directory and environment